Separate GET_/SET_/PATH cases in switch
This breaks out the 3 types of ET_PATH into their own case statements
in the switch.  It also adds an error when you try to use a GET_PATH
on a function with refinements, e.g. ':append/only'

(The error is temporary, as that should actually work.  It should give you
a function back that behaves as the refined from.)