Const correctness for routines used with paths
Put in their own commit to keep them separate from the larger change.

Here we see an actual good example of a "surprise" in a mutable
parameter handling.  Modify_Block would take in a value to be inserted
either as itself or as a block.  If you passed in the block itself, the value
you passed in would be modified to be a copy of the part of the block
being inserted.  Another reason why const correctness is your friend...