Remove TOS in Do_Path, address parse stackoverflow
Every time you used a path access in a parse rule, it leaked a stack item.
So foo: [a: "a"] then running parse "aaaaa" [some foo/a] would leave 5
items on the stack, once for each time it passed foo/a.  Loop that a bit
and you have a stack overflow.

This is because Do_Path would leave its result on the top of the stack,
and there was no handling of the cases where this stack value was left.
What made it hard to see was the introduction of an abstraction which
sometimes could find a variable in memory with no need for evaluation,
and sometimes needed to do a path which might have to fabricate a
new value somewhere.  Then there is the question of the lifetime of
that fabricated value...but when using the stack there was no clear place
to put a "conditional drop stack", because the knowledge of whether a
push had been done or not was removed.

Here we patch the problem simply, by having the relevant abstraction
take a scratch location that the caller provides.  It would be a simpler
interface if the result was always written into the location so it would be
a "destination", as opposed to "maybe using it".  But in theory this will
perform a little better by not having to copy a value in cases where a
pointer to the desired value can be acquired.

Looking into this brought up some other issues.  The REBPVS (used
for holding state during a path evaluation) has a number of REBVAL
pointers in it, and it's a known dangerous thing to be holding onto a
value pointer across evaluations in Rebol...because if the series that
value lives in gets modified, it can go bad.  StableStack helped ensure
that the values for arguments to natives and the output result of Do()
doesn't have to worry about that instability, but Do_Path() likely still
has concerns in this area.

Modulo bugs in this change, it should improve the situation overall.