Clarify series alloc/free, don't use swap "trick"
This is a clarifying rewrite of series allocation and freeing.  The
main drive for which was to get rid of a "trick" that was used in
the implementation of Expand_Series.  This would allocate a
new REBSER header and then blit the contents of the new
header into the old header and free the new series.

The argument for the trick was that it kept "external" code from
being aware of the inside of the series header mechanics.  But
only two routines were being "protected" (Expand_Series and
Shrink_Series).  The upshot was that any tracking that was trying
to be done at the point of series header allocation and free
started being mixed up with a spurious Make_Series...when
series resizing is more sensibly kept as an operation that only
affects the data series containing elements.

Beyond being more sensible, it saves an alloc/free on each
resize, and makes leak tracking code easier by not getting
tricked by the series header swapping.

This eliminates the Prop_Series routine, which had another use
in t-struct.c which should be vetted to see if it works correctly.