REBOL_Types enum => Reb_Kind
Because using the type enum is better documentation of what is
being passed (and type checked), there are now more uses of the
enumerated type holding REB_BLOCK, REB_STRING, etc.  Hence
writing out `enum REBOL_Types type = REB_BLOCK;` is both
lengthy and somewhat confusing.

At first this was just a rename to Reb_Type, but that would tread on
a hopeful future move to rename DATATYPE! to TYPE!, plus hamper
the associated switch of Reb_Datatype => Reb_Type.  Realizing that
and thinking ahead to when the switch comes, it would mean the
VAL_TYPE macro would have to be renamed for talking about the
enumeration instead of the instance of a Rebol datatype value.  So
I picked a conveniently short four-letter word that could work as
VAL_KIND being a replacement for VAL_TYPE.

As renaming the enum was the only real agenda item, this does not
do that change.  But it did change the macro VAL_DATATYPE to
VAL_TYPE_KIND to start trying the name change on for clarification.