Adjsut Host_Error retrieval for GCC
The strerror() and strerror_r() routine history is a bit of a mess,
and the hostkit error API isn't making it any better.  To make a long
story short: although you pass in a buffer to GNU's implementation
of strerror_r(), it may not use the buffer you provide for the error.
It may return a pointer to some constant data and leave the buffer
alone.  This change accounts for that to prevent garbage error
messages.