Delete outdated Trap comments, move longjmp comment
The c-error.c file had some comments that were out of date
already, and will be even more out of date after the Error
Renaissance commit.  The only thing worth saving was a
remark about how longjmp can return an integer value (I
didn't know, I thought it was just 1 or 0) and that the integer
is only 32-bit on 64-bit platforms.