RAISE and PANIC psuedo-keywords
This very large commit is the "Error Renaissance".  It has
several aspects, the most visible of which is that there are two
keyword-like macros called `raise` and `panic` which can be
used to give the appearance of something syntactically like
a throw construct:

    if (!IS_OBJECT(val))
        raise Error_Invalid_Arg(val);

How they work is explained in the comments of sys-core.h.
But the gist is that the crashing or longjmp'ing is not done by
the `raise` or the `panic`, but by the error function.  However
the keyword saves some state that the error function reads,
for instance to know whether to do a raise or panic action.

By combining this with the __attribute((noreturn)) and the
__builtin_unreachable extensions to GCC and Clang, the
uglier tricks using DEAD_END and DEAD_END void are not
necessary.  And the keyword-like syntax serves as a reminder
that something more is going on than just a function call--
that the next line of code will not be executed.

Doing this required unifying the RP_ Rebol Panic codes and
the RE_ Rebol Error codes more clearly.  So there's work in
here that does that, and removes some redundancy of
information that existed between strings.r and errors.r.  The
goal of Panic() is articulated more clearly and it actually acts
more like it seems to have set out to be.

Because every line with an error call on it needed to be
touched, I went ahead and did a light review of all the callsites
and cleaned things up a little bit as I went along.  This meant
shuffling around some of the functions that create errors when
I noticed a number of similar calls, and also removing
DEAD_END or break; instances that were right after a raise
or panic.  Some code formatting adjustments on the lines
above or below an error instantiation.