Make UNTIL arity-2, add LOOP-UNTIL & LOOP-WHILE
This transitions UNTIL to be arity-2, to be a natural pairing with the
WHILE construct.  (In the same spirit of IF vs. UNLESS).  The previous
version that is a condition only is then moved to LOOP-UNTIL, as a
way of saying what to do in the missing arity slot of the body.  Then
LOOP-WHILE is added to be the arity-1 WHILE complement of that.

Compatibility in <r3-legacy> is accomplished with:

    UNTIL: :LOOP-UNTIL

The word LOOP was chosen due to its generic sense.  ("What should
I be doing UNTIL that happens...or WHILE it is happening..." => "You
should loop.")  That's as opposed to having any specific relationship to
LOOP the construct's parameterization (now or under future redesign).

LOOP-WHILE has slightly more overhead than LOOP-UNTIL, because
it seeks to return the last TRUE? result (instead of being guaranteed to
return the FALSE or NONE value that caused the loop to break).  Thus
LOOP-UNTIL remains the "most efficient" base looping construct.