New REBFRM model + type incompatible with REBSER
The death of the FRAME! datatype was helpful in terminology, since
a "FRAME" was more generally a series which represented the key
and value pairings inside an object (or other context-like value).  This
series was implemented as two distinct series, which were historically
known as the "words" and "values".

This drives forward a new set of invariants, asserts, names, and rules
for a FRAME series...and also gives them their own datatype.  This
means that a REBFRM* and a REBSER* cannot be passed as equivalent
without some kind of cast.

That change alone meant having to touch a lot of code (that would not
compile until the change was pushed through).  But also, this tries to
nail down some other terms.  Because the "word list" does not contain
words (it contains typesets and a cached symbol) the name "keylist"
was deemed more appropriate.  And since "value" is very generic, the
idea of having a name for the corresponding slot to a key which was
a little different was good... the name "var" had been already adopted by
routines like `Get_Var` which would find the value slot in a frame for a
given word.

The routines for accessing keys and vars were simplified, assertions
strengthened in the debug build, and some general interface cleanup.
During debugging this required going in and handling some routines
that have changes which would be difficult to separate out in light of all
the other changes, so this includes those as well...with the hope that the
increase in asserts and correctness will make finding any bugs relatively
easy regardless.

This includes one functionality change made while fixing up the handling
of SET with objects.  Values that have been treated with PROTECT/HIDE
do not prevent a set, they merely are removed from the consideration:

    >> obj: object [a: 0 b: 0 c: 0]
    >> protect 'obj/b
    >> set obj [3 4]
    >> obj
    [a: 3 c: 4]

This should be helpful when SELF: is migrated to being a hidden <local>
of the object, similar to RETURN: in definitional scoped functions.