Restores some authorship information.
During the bulk conversion of file headers some authorship information was removed from file headers on the basis that it would be contained in a Credits file instead. The Credits file should have been created during the conversion process but was not because it was seen to be in a draft form. So since the bulk conversion that authorship information has been missing from Ren-C.  While Credits.md may not be finished, the new file headers reference it, it therefore should exist.

The authorship information in question can be viewed at: https://github.com/metaeducation/ren-c/blob/91f0cff531d3bc6b664d83cd5b812d72c0812e92/source-tool.issues.txt

As the file header bulk conversion was from my PR, I feel it necessary to submit this Credits.md to honor the information that was affected.