MSVC 2015 Zero-Warning CMake Build
This makes some changes to the build process, so that the MSVC 2015
build works and can be done without warnings.

Two warnings (which are actually the same warning, for assignment vs.
initialization) had to be disabled, because they were pervasively used.
They pertain to assigning value with more bits (e.g. 64-bit) to one
with less (e.g. 32-bit), which could mean a possible loss of data.
Most of these cases can be remedied if series indices are changed to
be 64-bit on 64-bit builds, which is a good idea and should be done.