Byte-Order Sensitive Bit Flags & Masking
All things being equal in the external C API, it would be desirable to
variadically mix strings with Rebol values, without pre-declaring
which is which.

In pseudocode matching libRed's "look" in the recent published code:

    reb_value block = rebLoad("[a b c]");
    reb_value date = rebLoad("12-Dec-2012");
    rebDo("poke", block, "2", date);

But though C has variadic functions (like printf), the callee receives
no type information from the callsite.  Before asking for each argument
the variadic function must know what type to ask for.

(Note: printf works around this by always taking one `char*` format
string as a first argument.  By analyzing the format string it then
"knows" the number and types of arguments to ask for.  Of course that
is risky, because the caller might get it wrong.  :-/)

One workaround for mixing string pointers and Rebol value pointers
would be if the variadic function asked for `void*` parameters...and
could then "sniff" each pointer to tell if it were a string or not.  If
the string literals are assumed to be in UTF-8, this is technically
possible... *if* the Rebol value headers are stylized to always look
like invalid UTF-8 strings.

It's made complicated because Rebol value headers want to be able to be
initialized, masked, and shifted like unsigned platform-natural
integers.  But since integers are mapped into memory differently by
different compilers on different architectures, getting the proper
imprint to register as "invalid UTF-8 string" has no good way that
is standards and/or performant.

This introduces macros for "Byte-Order Sensitive Bit Flags & Masking"
which try to address the problem.  The commit gets the bit positions
under control for an initial test, without changing the meanings of
the bits.  Once that's settled, selecting the bits so that Rebol
value headers are always invalid ways to begin a UTF-8 string will
be tackled.