Retake "writable if CPP" flag with NOT_FREE_MASK
Previously: when the Reb_Header of a series node was all 0 bits, this
was interpreted as a free node.  In order to give unmanaged series some
bit to set to make them look "not free", the NOT_END_MASK was used.
This prevented a REBSER header from being able to do double-duty as
a series END--but no need for that had (yet) arisen.

With the desire to make the leading two bits of most Reb_Headers
"10", this makes that leftmost bit mean NOT_FREE_MASK.  The bit is
also taken over as the sanity check that a REBVAL cell is not
uninitialized memory (with a ~50% probability of catching failed
cell initializations).

By combining the flag's purpose, it means that valid initialized
values cannot be marked unwritable.  This has little impact, as the
feature was (a) debug only, (b) only in the C++ build, (c) had only
been applied to a handful of canonized values (for END, TRUE, FALSE,
BLANK, etc.)  However, they are protected by macros which only expose
them as `const` most of the time...and there are benefits to not
having the stray bit on them so these canon values are safe to copy.
In fact, that bit had caused more problems than the writability check
had ever caught.  :-/

This step doesn't solve the problem of distinguishing arbitrary REBVAL
from UTF8.  But it does distinguish arbitrary *REBSER* from UTF8.  (It
is being done in an isolated commit to help with regression testing
from each step.)