First try at RIGOROUS warning settings for GCC
When Ren-C was started ("Coherence One") the goal was to make it
compile without warnings in C, C99, C++, C++11, C++14, etc. with
those levels turned up pretty high.  Then, as a practice, to treat
any warnings in the builds as errors.

While that got fairly far at the time, it was not completed, and the
switch to higher-than-default warning levels with warnings as errors
never did get done.  With the RIGOROUS setting, this attempts to
start putting that into practice.  It also provides a way to put in
per-file exemptions, which is particularly useful when compiling code
that did not originate from this project and shouldn't be edited
directly to fix pedantic errors--due to going out of sync being a
bad idea vs. sending patches to the official source.

Settings used are:

"-Werror" ;-- convert warnings to errors

"--pedantic" ;-- used only if not C89, due to "//" comments

"-Wextra"
"-Wall"

"-Wchar-subscripts"
"-Wwrite-strings"
"-Wundef"
"-Wformat=2"
"-Wdisabled-optimization"
"-Wlogical-op"
"-Wredundant-decls"
"-Woverflow"
"-Wpointer-arith"
"-Wparentheses"
"-Wmain"
"-Wsign-compare"
"-Wtype-limits"

When constness is being deliberately cast away, `m_cast` is used
(for "m"utability).  However, this is just a plain cast in C as it has
no const_cast.  Since the C language has no way to say you're doing a
mutability cast on purpose, the warning can't be used there... but
assume the C++ build covers it.

"-Wcast-qual" ; C++ only

Unfortunately, the majority of Rebol's C code was written with little
attention to overflow in arithmetic.  Frequently REBUPT is assigned to
REBCNT, size_t to REBYTE, etc.  The issue needs systemic review, but
will be most easy to do so when the core is broken out fully from less
critical code in optional extensions.

; DISABLED "-Wconversion"
; DISABLED "-Wstrict-overflow"